feat: add accounts from account list in AccountId by default
